Several bugfixes, resulting version seems quite (st)able
- check that nodes in copied / moved tree are not the dest node (or any of
its parents). Consider the following case

Base    Edited
   R     R
  a b   a b
  c     c
        b
Now we might decide to move/copy the tree c to b. But the subtree c has
b as a child! => Trouble. The bugfix (_treeHasNoCopyOf) check for this

- fixes to doGetRandomNode to make sure we don't get into the forbidden
tree and/or recurse with null.

- ALSO: outcommented code for dividing edits into dium and c phases -- this
might be a good idea if further copy-related bugs occur